Update debug toolbar and move monkey patch code into model files to fix app startup issues
