+++ pip-freeze.txt
@@ -16,7 +16,7 @@ django-appconf==1.0.1
 django-celery==3.1.17
 django-celery-email==1.1.4
 django-compressor==1.6
-django-debug-toolbar==1.3.2
+django-debug-toolbar==1.5
 django-digest==1.13
 django-el-pagination==2.1.1
 django-guardian==1.3.2
